Added DTI class and tests
