+++ ukat/mapping/__init__.py
@@ -1,1 +1,1 @@
-from . import b0, dwi, t1, t2, t2star
+from . import b0, diffusion, t1, t2, t2star
