+++ ukat/mapping/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_b0, test_dwi, test_t1, test_t2, test_t2star
+from . import test_b0, test_diffusion, test_t1, test_t2, test_t2star
